Added Log in (Jira TA-3)
