--- src/main/java/org/loose/fis/registration/example/controllers/RegistrationController.java
@@ -1,40 +1,0 @@
-package org.loose.fis.registration.example.controllers;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.ChoiceBox;
-import javafx.scene.control.PasswordField;
-import javafx.scene.control.TextField;
-import javafx.scene.text.Text;
-import org.loose.fis.registration.example.exceptions.UsernameAlreadyExistsException;
-import org.loose.fis.registration.example.services.UserService;
-
-public class RegistrationController {
-
-    @FXML
-    private Text registrationMessage;
-    @FXML
-    private PasswordField passwordField;
-    @FXML
-    private TextField usernameField;
-    @FXML
-    private TextField email_addressField;
-    @FXML
-    private TextField phone_numberField;
-    @FXML
-    private ChoiceBox role;
-
-    @FXML
-    public void initialize() {
-        role.getItems().addAll("Learner", "Tutor");
-    }
-
-    @FXML
-    public void handleRegisterAction() {
-        try {
-            UserService.addUser(usernameField.getText(), email_addressField.getText(), phone_numberField.getText(), passwordField.getText(), (String) role.getValue());
-            registrationMessage.setText("The account was created successfully! ^_^");
-        } catch (UsernameAlreadyExistsException e) {
-            registrationMessage.setText(e.getMessage());
-        }
-    }
-}
