--- src/main/java/org/loose/fis/registration/example/exceptions/CouldNotWriteUsersException.java
@@ -1,4 +1,0 @@
-package org.loose.fis.registration.example.exceptions;
-
-public class CouldNotWriteUsersException extends RuntimeException {
-}
