--- src/main/java/org/loose/fis/registration/example/exceptions/UsernameAlreadyExistsException.java
@@ -1,15 +1,0 @@
-package org.loose.fis.registration.example.exceptions;
-
-public class UsernameAlreadyExistsException extends Exception {
-
-    private String username;
-
-    public UsernameAlreadyExistsException(String username) {
-        super(String.format("An account with the username %s already exists!", username));
-        this.username = username;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-}
