+++ src/main/resources/register.fxml
@@ -6,19 +6,11 @@
 <?import javafx.scene.text.Text?>
 <GridPane xmlns:fx="http://javafx.com/fxml/1" alignment="center" hgap="10" prefHeight="400" prefWidth="600"
           vgap="10" xmlns="http://javafx.com/javafx/10.0.2-internal"
-<<<<<<< HEAD
           fx:controller="TA.controllers.Controller">
     <padding>
         <Insets bottom="25.0" left="25.0" right="25.0" top="25.0"/>
     </padding>
     <Text id="title-text" text="Register" GridPane.columnIndex="0" GridPane.columnSpan="2"
-=======
-          fx:controller="org.loose.fis.registration.example.controllers.RegistrationController">
-    <padding>
-        <Insets bottom="25.0" left="25.0" right="25.0" top="25.0"/>
-    </padding>
-    <Text id="title-text" text="Tutoring Application" GridPane.columnIndex="0" GridPane.columnSpan="2"
->>>>>>> 9fe96ed288e82ecb9bcc62270b1674b669014667
           GridPane.rowIndex="0"/>
 
     <Label text="Username:" GridPane.columnIndex="0" GridPane.rowIndex="1"/>
@@ -44,10 +36,6 @@
 
     <Text fx:id="registrationMessage" GridPane.columnIndex="1" GridPane.rowIndex="8"/>
     <HBox alignment="bottom_right" spacing="10" GridPane.columnIndex="1" GridPane.rowIndex="7">
-<<<<<<< HEAD
         <Button onAction="#startRegister" text="Register"/>
-=======
-        <Button onAction="#handleRegisterAction" text="Register"/>
->>>>>>> 9fe96ed288e82ecb9bcc62270b1674b669014667
     </HBox>
 </GridPane>
